Implemented getParkingSpaceById
In ParkingSpaceProductionImpl. Also renamed the class so it's easier to refer to

+++ src/main/java/com/airpnp/service/ParkingSpaceMockImpl.java
@@ -33,7 +33,7 @@ public class ParkingSpaceMockImpl implements ParkingSpaceService {
     }
 
     @Override
-    public ParkingSpace getParkingSpaceById(int Id) {
+    public ParkingSpace getParkingSpaceById(Long Id) {
         return parkingSpaces.get(Id);
     }
